requirements.txt
Django
psycopg2-binary
pillow
djangorestframework
django-filter
twilio